Move ValuesForTesting into test directory
