+++ test/util/IdTableHelpers.cpp
@@ -4,12 +4,10 @@
 
 #include "../test/util/IdTableHelpers.h"
 
-#include <absl/strings/str_cat.h>
-
 #include <algorithm>
 #include <utility>
 
-#include "engine/ValuesForTesting.h"
+#include "../engine/ValuesForTesting.h"
 #include "engine/idTable/IdTable.h"
 #include "global/ValueId.h"
 #include "util/Algorithm.h"
